+++ objects/geometry_objects/Sphere.h
@@ -8,14 +8,14 @@ namespace objects {
 
 class Sphere: public GeometryObject {
 public:
-    Sphere(Vector3d position, int radius);
+    Sphere(Vector3d position, double radius);
 
     void intersectRay(Ray &ray, double &int1, double &int2, bool &isInt) override;
     void transform(std::shared_ptr<math::Matrix> matrix) override;
     Vector3d getNormal(Vector3d &point, const Vector3d &direction) const override;
 private:
     Vector3d _position;
-    int _radius;
+    double _radius;
 };
 
 } // namespace objects
